requirements.txt
pandas==2.2.2
numpy>=1.23.5,<2.0.0
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.3.2
tensorflow-macos==2.15.0
tensorflow-metal==1.2.0
eli5==0.13.0
scipy==1.13.0